refactor: fix newly-introduced linting issues
